apps/cognee_service/requirements.txt
# NOEMA Cognee Service Dependencies
# Cognee provides semantic memory infrastructure

# Core Cognee
cognee>=0.1.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Data validation
pydantic>=2.5.0

# Environment management
python-dotenv>=1.0.0

# HTTP client (for health checks)
httpx>=0.26.0
